Conflicts:
	karbon/data/karbon.appdata.xml
	krita/data/krita_filefilter.desktop
	krita/plugins/extensions/layersplit/kritalayersplit.desktop
	krita/ui/forms/wdggeneralsettings.ui
	plugins/reporting/chart/koreport_chartplugin.desktop
	plugins/reporting/web/koreport_webplugin.desktop
	stage/templates/odf/cohere-blue.desktop
	stage/templates/odf/cohere-gold.desktop
	stage/templates/odf/cohere-magenta.desktop
	stage/templates/odf/cohere-red.desktop

+++ krita/image/kis_rect_mask_generator.cpp
@@ -72,7 +72,7 @@ void KisRectangleMaskGenerator::setScale(qreal scaleX, qreal scaleY)
 void KisRectangleMaskGenerator::setSoftness(qreal softness)
 {
     KisMaskGenerator::setSoftness(softness);
-    qreal safeSoftnessCoeff = 1.0 / qMax(0.01, softness);
+    qreal safeSoftnessCoeff = qreal(1.0) / qMax(qreal(0.01), softness);
 
     d->transformedFadeX = d->xfadecoeff * safeSoftnessCoeff;
     d->transformedFadeY = d->yfadecoeff * safeSoftnessCoeff;
